+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Revision history for json-syntax
 
-## 0.1.3.0 -- 2021-??-??
+## 0.2.0.0 -- 2021-??-??
 
+* Switch from `Chunks` to `SmallArray` in the `Object` and `Array` data
+  constructors. This makes the library simpler to use but it a breaking
+  change.
 * Add `object(9|10|11|12)` as convenience helpers for construction.
 
 ## 0.1.2.0 -- 2020-11-18